Truncate the deltas file instead of deleting it.
git-svn-id: https://msnp-sharp.googlecode.com/svn/trunk/MSNP-Sharp15@663 ed960e22-f52a-0410-8d30-b9dff305f3af

+++ MSNPSharp/ContactService.cs
@@ -1976,8 +1976,15 @@ namespace MSNPSharp
                 string deltasResultFile = Path.Combine(Settings.SavePath, NSMessageHandler.Owner.Mail.GetHashCode() + "d" + ".mcl");
                 if (File.Exists(deltasResultFile))
                 {
-                    File.SetAttributes(deltasResultFile, FileAttributes.Normal);  //By default, the file is hidden.
-                    File.Delete(deltasResultFile);
+                    if (Deltas != null)
+                    {
+                        Deltas.Truncate();  //If we saved cachekey and preferred host in it, deltas can't be deleted.
+                    }
+                    else
+                    {
+                        File.SetAttributes(deltasResultFile, FileAttributes.Normal);  //By default, the file is hidden.
+                        File.Delete(deltasResultFile);
+                    }
                 }
                 abSynchronized = false;
             }
